[server] | refactor(): 공통 모듈 경로 이동
