+++ packages/server/src/common/errors.ts
